Improve Navigation buttons, properly display them on mobile
